backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
email-validator
passlib
bcrypt
PyJWT
python-multipart
google-generativeai
openai
python-dotenv
